Issue #13: Refactoring for Delegate
